requirements.txt
Django==2.2.4
django-bootstrap3==11.1.0
django-filter==2.2.0
django-model-utils
Pillow
